--- 6_player_timer.cpp
@@ -1,11 +1,0 @@
-#include <stdio.h>
-#include "pico/stdlib.h"
-
-int main()
-{
-    stdio_init_all();
-
-    puts("Hello, world!");
-
-    return 0;
-}
