+++ CMakeLists.txt
@@ -5,10 +5,6 @@ cmake_minimum_required(VERSION 3.13)
 set(CMAKE_C_STANDARD 11)
 set(CMAKE_CXX_STANDARD 17)
 
-# Initialise pico_sdk from installed location
-# (note this can come from environment, CMake cache etc)
-set(PICO_SDK_PATH CACHE PATH "/usr/lib/pico-sdk")
-
 # Set board to pico
 set(PICO_BOARD pico CACHE STRING "Board type")
 
@@ -25,23 +21,20 @@ project(6_player_timer C CXX ASM)
 pico_sdk_init()
 
 # Add executable. Default name is the project name, version 0.1
-add_executable(6_player_timer 6_player_timer.cpp)
+add_executable(6_player_timer)
+
+# Add subdirectories
+add_subdirectory(src)
 
 # Add the standard library to the build
 target_link_libraries(6_player_timer pico_stdlib)
+target_link_libraries(6_player_timer timer_components)
 
 # Add the standard include files to the build
 target_include_directories(6_player_timer PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}
-  ${CMAKE_CURRENT_LIST_DIR}/include
   ${CMAKE_CURRENT_LIST_DIR}/.. # for our common lwipopts or any other standard includes, if required
 )
-
-# Add main
-add_subdirectory(src)
-
-# Add main source
-add_subdirectory(test)
 
 # Set project name and version
 pico_set_program_name(6_player_timer "6_player_timer")
