--- include/buzzer.hpp
@@ -1,6 +1,0 @@
-class Buzzer {
-    const int buzzer_pin;
-public:
-    Buzzer(int new_buzzer_pin);
-    void beep();
-};
