+++ include/gont/pico/CmakeLists.txt
@@ -1,0 +1,1 @@
+target_include_directories(timer_components INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
