+++ src/CMakeLists.txt
@@ -1,1 +1,3 @@
-target_sources(6_player_timer PRIVATE buzzer.cpp)
+target_sources(6_player_timer PUBLIC 6_player_timer.cpp)
+
+add_subdirectory(gont/pico)
