+++ src/gont/pico/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(timer_components buzzer.cpp)
+
+target_include_directories(timer_components INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
