+++ src/gont/pico/buzzer.hpp
@@ -1,0 +1,6 @@
+class Buzzer {
+  const int buzzer_pin;
+public:
+  Buzzer(int new_buzzer_pin);
+  void beep();
+};
